# Conflicts:
#	zzic-back/src/main/java/point/zzicback/mapper/TodoMapper.xml
#	zzic-back/src/main/java/point/zzicback/service/TodoService.java

+++ zzic-back/src/main/java/point/zzicback/service/TodoServiceImpl.java
@@ -19,7 +19,7 @@ public class TodoServiceImpl implements TodoService {
     }
 
     @Override
-    public Todo getTodoById(Integer id) {
+    public Todo getTodoById(Long id) {
         // 임시로 null 반환
         return this.todoMapper.selectByPrimaryKey(id.longValue());
     }
@@ -35,7 +35,7 @@ public class TodoServiceImpl implements TodoService {
     }
 
     @Override
-    public void deleteTodo(Integer id) {
+    public void deleteTodo(Long id) {
         // 임시로 아무 작업도 하지 않음
     }
 }
